Make bootloader work with other flash page sizes
Previously it only worked with 1 KiB flash pages.
We no longer copy the file contents to a predefined address.
We now rely on the FAT to be ordered (i.e. consecutive sectors are one after each other in memory) after the file is copied.
So we now only need to jump to the start of the first sector (0x8800 for now).